move content of loginPage and startPage to new PageObject structure
-move /pages/loginPage.js content to  pages/generalPagesBeforeLogin/LoginPage.js
-move /pages/startPage.js content to  pages/generalPagesBeforeLogin/StartPageBeforeLogin.js
-change links  to those
- delete:  /pages/loginPage.js, /pages/startPage.js
