+++ page-objects/pages/generalPagesBeforeLogin/StartPageBeforeLogin.js
@@ -1,0 +1,14 @@
+'use strict';
+
+module.exports = {
+
+    selectors: {
+        frontpageLoginBtn: 'a[data-testid="login-btn"]',
+
+    },
+
+    clickLoginBtn: async function () {
+        let frontpageLoginBtn = await driver.$(this.selectors.frontpageLoginBtn);
+        await frontpageLoginBtn.click();
+    }
+}
