+++ shared-objects/loginData.js
@@ -15,7 +15,7 @@ module.exports = {
   defaultAdminPassword: 'Schulcloud1!',
 
   elem: {
-    frontpageLoginBtn: 'a[data-testid="login-btn"]',
+    //frontpageLoginBtn: 'a[data-testid="login-btn"]',
     usernameInput: 'section#loginarea input[data-testid="username"]',
     passwordInput: 'section#loginarea input[data-testid="password"]',
     loginBtn: 'section#loginarea input[data-testid="submit-login"]',
