+++ step_definitions/addPupilsToTheCourse-steps.js
@@ -1,8 +1,8 @@
 'use strict';
 
 const addPupilToTheCourse = require('../page-objects/addPupilToTheCourse');
-const loginPage = require('../page-objects/pages/loginPage');
-const startPage = require('../page-objects/pages/startPage');
+const loginPage = require('../page-objects/pages/generalPagesBeforeLogin/LoginPage.js');
+const startPage = require('../page-objects/pages/generalPagesBeforeLogin/StartPageBeforeLogin.js');
 const createCourse = require('../page-objects/createCourse');
 const Login = require('../shared-objects/loginData');
 const elementHelpers = require('../runtime/helpers/elementHelpers.js');
