+++ step_definitions/administartion-steps.js
@@ -4,8 +4,8 @@ let addPupilToTheCourse = require('../page-objects/addPupilToTheCourse');
 let administration = require('../page-objects/administration');
 //let teacherLogin = require('../page-objects/teacherLogin');
 const elementHelpers = require('../runtime/helpers/elementHelpers.js')
-const loginPage = require('../page-objects/pages/loginPage');
-const startPage = require('../page-objects/pages/startPage');
+const loginPage = require('../page-objects/pages/generalPagesBeforeLogin/LoginPage.js');
+const startPage = require('../page-objects/page-objects/pages/generalPagesBeforeLogin/StartPageBeforeLogin.js');
 const copyCourse = require('../page-objects/copyCourse');
 const loginData = require('../shared-objects/loginData');
 const courseData = require('../shared-objects/courseData');
