+++ step_definitions/createTask-steps.js
@@ -3,8 +3,8 @@
 const createTask = require("../page-objects/createTask");
 const loginData = require("../shared-objects/loginData");
 const firstLogin = require("../shared_steps/firstLogin.js");
-const loginPage = require('../page-objects/pages/loginPage');
-const startPage = require('../page-objects/pages/startPage');
+const loginPage = require('../page-objects/pages/generalPagesBeforeLogin/LoginPage.js');
+const startPage = require('../page-objects/page-objects/pages/generalPagesBeforeLogin/StartPageBeforeLogin.js');
 const elementHelpers = require('../runtime/helpers/elementHelpers.js');
 let page = { createTask };
