+++ step_definitions/wrongPasswordLogin-steps.js
@@ -1,8 +1,8 @@
 'use strict';
 
 const elementHelpers = require('../runtime/helpers/elementHelpers.js');
-const loginPage = require('../page-objects/pages/loginPage');
-const startPage = require('../page-objects/pages/startPage');
+const loginPage = require('../page-objects/pages/generalPagesBeforeLogin/LoginPage.js');
+const startPage = require('../page-objects/page-objects/pages/generalPagesBeforeLogin/StartPageBeforeLogin.js');
 let loginData = require('../shared-objects/loginData');
 let shared = { loginData };
